Seperating Telemetry behaviour from feedback.
the goal of this is to seperate telemetry functions from the display functions.
This means that recieveing telemetry is no longer dependant on showing the "home screen" and the "home screen" can be updated without recieveing new telemetry data.

This alows for future altering of whats being displayed whilst lowering the risk of affecting remote operation